[ADD] Practica 1

+++ Practicas/Practica_1/ej7.py
@@ -1,7 +1,7 @@
 
-horas = int(input("Introduce las Horas Trabajadas: "))
-precio_hora = int(input("Introduce el Precio de la Hora: "))
-retencion = int(input("Introduce la retencion aplicable en %: "))
+horas = float(input("Introduce las Horas Trabajadas: "))
+precio_hora = float(input("Introduce el Precio de la Hora: "))
+retencion = float(input("Introduce la retencion aplicable en %: "))
 
 salario_b = horas * precio_hora
 salario_n = salario_b - (salario_b * (retencion/100))
